# public/swagger.json
{
	"components": {
...
		"schemas": {
			"IResponse": {
				"properties": {
					"error": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "boolean"
							}
						]
					},
					"message": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ISuccessResponse": {
				"properties": {
					"error": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "boolean"
							}
						]
					},
					"message": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"ISignInResponse": {
				"properties": {
					"access_token": {
						"type": "string",
						"nullable": true
					},
					"refresh_token": {
						"type": "string",
						"nullable": true
					},
					"error": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"access_token",
					"refresh_token"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IAuthAccessTokenResp": {
				"properties": {
					"access_token": {
						"type": "string"
					}
				},
				"required": [
					"access_token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAuthRefreshToken": {
				"properties": {
					"refresh_token": {
						"type": "string"
					}
				},
				"required": [
					"refresh_token"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IToken": {
				"properties": {
					"user_id": {
						"type": "string"
					},
					"access_token": {
						"type": "string",
						"nullable": true
					},
					"access_token_expires_at": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"refresh_token": {
						"type": "string",
						"nullable": true
					},
					"refresh_token_expires_at": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"user_id",
					"access_token",
					"access_token_expires_at",
					"refresh_token",
					"refresh_token_expires_at",
					"description",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITokenInsert": {
				"properties": {
					"user_id": {
						"type": "string"
					},
					"access_token": {
						"type": "string",
						"nullable": true
					},
					"access_token_expires_at": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"refresh_token": {
						"type": "string",
						"nullable": true
					},
					"refresh_token_expires_at": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"description": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"user_id",
					"access_token",
					"access_token_expires_at",
					"refresh_token",
					"refresh_token_expires_at",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITokenUpdate": {
				"properties": {
					"id": {
						"type": "string"
					},
					"access_token": {
						"type": "string",
						"nullable": true
					},
					"access_token_expires_at": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"refresh_token": {
						"type": "string",
						"nullable": true
					},
					"refresh_token_expires_at": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"description": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUser": {
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password_salt": {
						"type": "string"
					},
					"is_sso_user": {
						"type": "boolean"
					},
					"sso_user_id": {
						"type": "string",
						"nullable": true
					},
					"sso_from": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"username",
					"password",
					"email",
					"password_salt",
					"is_sso_user",
					"sso_user_id",
					"sso_from",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserInsert": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password_salt": {
						"type": "string",
						"nullable": true
					},
					"is_sso_user": {
						"type": "boolean"
					},
					"sso_user_id": {
						"type": "string",
						"nullable": true
					},
					"sso_from": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserUpdate": {
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password_salt": {
						"type": "string"
					},
					"is_sso_user": {
						"type": "boolean"
					},
					"sso_user_id": {
						"type": "string",
						"nullable": true
					},
					"sso_from": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "express-authen-ts",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/v1/auth/refresh/token": {
			"post": {
				"operationId": "AuthRefreshToken",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/IAuthAccessTokenResp"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"access_token": "new_access_token"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAuthRefreshToken"
							}
						}
					}
				}
			}
		},
		"/v1/auth/refresh/tokens": {
			"post": {
				"operationId": "AuthBothTokens",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/ISignInResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"access_token": "new_access_token",
											"refresh_token": "new_refresh_token"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAuthRefreshToken"
							}
						}
					}
				}
			}
		},
		"/v1/revoke/token/{user_id}": {
			"post": {
				"operationId": "RevokeToken",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/ISuccessResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "success"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "user_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/token/{id}": {
			"get": {
				"operationId": "GetToken",
				"responses": {
					"200": {
						"description": "Get token",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/IToken"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": "5",
											"user_id": "test",
											"access_token": "access_token",
											"access_token_expires_at": 1660926192000,
											"refresh_token": "refresh_token",
											"refresh_token_expires_at": 1660926192000,
											"description": "secret"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateToken",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/ISuccessResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Successfully update",
											"id": "123"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ITokenUpdate"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteToken",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/ISuccessResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Successfully delete",
											"id": "123"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/user/tokens/{user_id}": {
			"get": {
				"operationId": "GetTokenByUserId",
				"responses": {
					"200": {
						"description": "Get tokens",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/IToken"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"id": "5",
												"user_id": "test",
												"access_token": "access_token",
												"access_token_expires_at": 1660926192000,
												"refresh_token": "refresh_token",
												"refresh_token_expires_at": 1660926192000,
												"description": "secret"
											}
										]
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "user_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/token": {
			"post": {
				"operationId": "AddToken",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/ISuccessResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Successfully create",
											"id": "123"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ITokenInsert"
							}
						}
					}
				}
			}
		},
...
		"/v1/user/{id}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Get user",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/IUser"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": "3",
											"username": "test2",
											"password": "test1234",
											"password_salt": "test",
											"email": "john2@email.com",
											"is_sso_user": false,
											"sso_user_id": null,
											"sso_from": null,
											"status": "active"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/ISuccessResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Successfully update",
											"id": "123"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserUpdate"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/ISuccessResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Successfully delete",
											"id": "123"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/user": {
			"post": {
				"operationId": "AddUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IResponse"
										},
										{
											"$ref": "#/components/schemas/ISuccessResponse"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Successfully create",
											"id": "123"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserInsert"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}